+++ src/euroeval/tasks.py
@@ -350,8 +350,7 @@ MULTIPLE_CHOICE = Task(
     metrics=[m.mcc_metric, m.accuracy_metric],
     default_num_few_shot_examples=5,
     default_max_generated_tokens=NUM_GENERATION_TOKENS_FOR_CLASSIFICATION,
-    default_labels=None,
-    default_allowed_model_types=[ModelType.GENERATIVE],
+    default_labels=None, # TODO (DONE): Add the Encoder ModelType
     uses_logprobs=True,
 )